+++ package.json
@@ -14,6 +14,8 @@
     "author": "",
     "license": "ISC",
     "dependencies": {
+        "@emotion/react": "^11.11.4",
+        "@emotion/styled": "^11.11.5",
         "@ijl/cli": "^5.1.0",
         "@types/react": "^18.2.74",
         "@types/react-dom": "^18.2.24",
