+++ src/app.tsx
@@ -1,15 +1,15 @@
-import React from 'react';
-import U2k from './4u2k_page'
+import React from 'react'
+import U2k from './pages/4u2k_page'
 import B2B from './b2b_page'
 import Main from './main_page'
 import ABC from './abc'
 import DoUHearMe from './douhearme'
 import Games from './games'
-import Price from './price_page'
+import Price from './pages/price_page'
 const App = () => {
-  return (
-    <Main />
-  );
+    return (
+        <B2B />
+    );
 };
 
 export default App;
