+++ src/components/header/header.styled.ts
@@ -1,0 +1,15 @@
+import styled from "@emotion/styled"
+import { css } from '@emotion/react'
+
+export const HeaderStyle = styled.header`
+    padding: 0px 96px 0px 96px;
+    display: flex;
+    flex-wrap: wrap;
+    flex-direction: row;
+    justify-content: space-between;
+    background-color: #D9D9D9;
+    align-items: center;
+    font-size: 2vw;
+    height: 8vw;
+
+`;
