+++ src/components/header/header.tsx
@@ -1,0 +1,18 @@
+import React from "react"
+import
+
+    < header className = "header" >
+                    <h1 className="logo">
+                        <span className="red_letter">e</span>sc
+                    </h1>
+                    <nav>
+                        <ul className="header-nav-menu">
+                            <li><a href="main_page.html">home</a></li>
+                            <li><a href="price_page.html">prices</a></li>
+                            <li><a href="4u2k_page.html">4u2k</a></li>
+                            <li><a href="games_page.html">games</a></li>
+                            <li><a href="b2b_page.html">b2b</a></li>
+
+                        </ul>
+                    </nav>
+                </header >
