+++ src/components/link/link.styled.ts
@@ -1,0 +1,7 @@
+import styled from "@emotion/styled"
+import { css } from '@emotion/react'
+
+export const StyledLink = styled.a`
+    text-decoration: none;
+    color: #000000
+`;
