+++ src/styles/main.css
@@ -1,0 +1,5 @@
+:root {
+    --text-highlight: #C00000;
+    --secondary-color: #D9D9D9;
+    --default-font-size: 1.5vw;
+}
